Library Management: add service interfaces
 * moved all service classes to implementation folder
 * created interfaces for each service
 * implemented polymorphism
