+++ LibraryProject/src/main/java/com/example/libraryproject/service/GoogleBooksApiService.java
@@ -1,0 +1,10 @@
+package com.example.libraryproject.service;
+
+import com.example.libraryproject.model.dto.BookAdditionFromGoogleRequest;
+
+public interface GoogleBooksApiService {
+
+    void fetchAndSaveBooks();
+
+    boolean fetchBook(BookAdditionFromGoogleRequest request);
+}
